backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Ксения Каткова",email = "ksusha.katkova2468@yandex.ru"}
]
requires-python = ">=3.14"
dependencies = [
    "alembic (>=1.18.4,<2.0.0)",
    "fastapi (>=0.129.0,<0.130.0)",
    "uvicorn (>=0.41.0,<0.42.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "python-multipart (>=0.0.22,<0.0.23)",
    "pandas (>=3.0.0,<4.0.0)",
    "pydantic-settings (>=2.13.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
